the code for the spin coater
Its a hard drive disc motor controlled with an esc. The LCD is I2C. The keypad is a basic 4x4. The IR sensor is interfaced with a little chip I cant remember the name of, but has 3 pins (ground vcc and output). I am trying to get the thing to calibrate the esc when it turns on and then get an input value from the keypad and display it on the first row of the LCD. It uses that input value as a setpoint for a PID loop that takes its input from a few functions that determine the rpm from a small part of the rotor that has no black nail polish covering it (its reflective metal). The PID output goes to control the PWM sent to the ESC. The rpm value is then read every second or more and displayed on the second row of the LCD.